Moved Vc helper math functions to a new file
